CIV-12358 - add unit test + refactored integration tests
