+++ dashboard-notifications/build.gradle
@@ -26,6 +26,7 @@ dependencies {
   implementation group: 'com.vladmihalcea', name: 'hibernate-types-52', version: '2.21.1'
   implementation group: 'io.swagger.core.v3', name: 'swagger-annotations', version: '2.2.9'
   testImplementation 'junit:junit:4.13.1'
+  testImplementation 'org.junit.jupiter:junit-jupiter:5.8.1'
   runtimeOnly 'org.postgresql:postgresql'
   compileOnly 'org.projectlombok:lombok'
   developmentOnly 'org.springframework.boot:spring-boot-devtools'
