+++ dashboard-notifications/src/main/java/uk.gov.hmcts.reform/dashboard/repositories/NotificationRepository.java
@@ -3,7 +3,6 @@ package uk.gov.hmcts.reform.dashboard.repositories;
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 import uk.gov.hmcts.reform.dashboard.entities.NotificationEntity;
-import java.util.Optional;
 import java.util.UUID;
 import javax.transaction.Transactional;
 
@@ -11,5 +10,4 @@ import javax.transaction.Transactional;
 @Repository
 public interface NotificationRepository extends CrudRepository<NotificationEntity, UUID> {
 
-    Optional<NotificationEntity> getNotification(UUID uuid);
 }
